+++ force-app/main/default/lwc/anonymiseRecord/anonymiseRecord.html
@@ -9,98 +9,23 @@
         <p>Some info</p>
       </lightning-tab>
 
-      <lightning-tab label="Users">
-        <div class="slds-m-bottom_medium">
-          <lightning-button
-            variant="base"
-            label="Add fields"
-            title="Destructive action"
-            onclick={handleClick}
-          ></lightning-button>
-        </div>
-
-        <div>
-          <lightning-dual-listbox
-            name="languages"
-            label="Please select fields"
-            source-label="Available"
-            selected-label="Selected"
-            field-level-help="Select fields to anonymise"
-            options={options}
-            required-options={requiredOptions}
-            value={values}
-            onchange={handleChange}
-          ></lightning-dual-listbox>
-        </div>
+      <lightning-tab label="User">
+        <c-field-selection-tab
+          objectname="User"
+          onfieldschange={handleFieldsChange}
+        ></c-field-selection-tab>
       </lightning-tab>
 
-      <lightning-tab label="Accounts">
-        <div class="slds-m-bottom_medium">
-          <lightning-button
-            variant="base"
-            label="Add fields"
-            title="Destructive action"
-            onclick={handleClick}
-          ></lightning-button>
-        </div>
-
-        <div>
-          <lightning-dual-listbox
-            name="languages"
-            label="Select Languages"
-            source-label="Available"
-            selected-label="Selected"
-            field-level-help="Select your preferred languages"
-            options={options}
-            onchange={handleChange}
-          ></lightning-dual-listbox>
-        </div>
+      <lightning-tab label="Account">
+        <c-field-selection-tab objectname="Account"></c-field-selection-tab>
       </lightning-tab>
 
-      <lightning-tab label="Contacts">
-        <div class="slds-m-bottom_medium">
-          <lightning-button
-            variant="base"
-            label="Add fields"
-            title="Destructive action"
-            onclick={handleClick}
-          ></lightning-button>
-        </div>
-
-        <div>
-          <lightning-dual-listbox
-            name="languages"
-            label="Select Languages"
-            source-label="Available"
-            selected-label="Selected"
-            field-level-help="Select your preferred languages"
-            options={options}
-            onchange={handleChange}
-          ></lightning-dual-listbox>
-        </div>
+      <lightning-tab label="Contact">
+        <c-field-selection-tab objectname="Contact"></c-field-selection-tab>
       </lightning-tab>
 
-      <lightning-tab label="Leads">
-        <div class="slds-m-bottom_medium">
-          <lightning-button
-            variant="base"
-            label="Add fields"
-            title="Destructive action"
-            onclick={handleClick}
-          ></lightning-button>
-        </div>
-
-        <div>
-          <lightning-dual-listbox
-            name="languages"
-            label="Select Languages"
-            source-label="Available"
-            selected-label="Selected"
-            field-level-help="Select your preferred languages"
-            options={options}
-            onchange={handleChange}
-          ></lightning-dual-listbox>
-        </div>
+      <lightning-tab label="Lead">
+        <c-field-selection-tab objectname="Lead"></c-field-selection-tab>
       </lightning-tab>
     </lightning-tabset>
 
@@ -110,7 +35,7 @@
       <h1>Currently selected fields:</h1>
 
       <div class="slds-box slds-theme_shade slds-m-top_medium">
-        <p>{fields}</p>
+        <p>{allFields}</p>
       </div>
     </div>
