+++ force-app/main/default/lwc/anonymiseRecord/anonymiseRecord.js
@@ -1,43 +1,13 @@
-import { LightningElement } from "lwc";
-import getFields from "@salesforce/apex/A2S_FieldSelectionHandler.getFields";
+import { LightningElement, track } from "lwc";
 
 export default class AnonymiseRecord extends LightningElement {
-  options = [];
-  values = [];
-  requiredOptions = [];
-  isFirst = true;
-  _fields = [];
+  @track fields = [];
 
-  async connectedCallback() {
-    await this.fetchData();
+  get allFields() {
+    return this.fields.length ? this.fields : "none";
   }
 
-  get fields() {
-    return this._fields.length ? this._fields : "";
-  }
-
-  handleChange(e) {
-    this._fields = e.detail.value;
-  }
-
-  async fetchData() {
-    await getFields({
-      objectname: "User"
-    })
-      .then((result) => {
-        let data = JSON.parse(JSON.stringify(result));
-        let lstOption = [];
-        for (let i = 0; i < data.length; i++) {
-          lstOption.push({
-            value: data[i].QualifiedApiName,
-            label: data[i].DeveloperName
-          });
-        }
-        this.options = lstOption;
-        console.log(lstOption);
-      })
-      .catch((error) => {
-        console.error(error);
-      });
+  handleFieldsChange(e) {
+    this.fields = e.detail;
   }
 }
